+++ CHANGELOG.md
@@ -6,7 +6,14 @@ This file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [2.5.1] - 2023-12-31
+## [2.5.2] - 2023-12-31
+
+### Fixed:
+  - Reverted "Added `const` to return generic types of `map`, `mapKeys`, `mapValues`, `flatMap` and `reduce`." from
+    previous version.  It was breaking in unpredicted ways that were way more impactful that I thought.
+
+
+## [2.5.1] - 2023-12-31 [YANKED]
 
 ### Changed:
   - Fixed documentation for `mapKeys` and `mapValues`